+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 
 - Log message when setting user before starting the SDK (#4882)
+- Add experimental flag to disable swizzling of `NSData` individually (#4859)
 
 ## 8.45.0